Add files via upload
The marccd.py is reading the marccd type images, with a function for reading the header. This takes a filename and returns a dictionary of all the information, and a function for reading the image, which also takes a filename and returns a two-dimensional array of all the points. 

The mask.py file takes an image (in the form of a two-dimensional array, and returns a two-dimensional array containing only 0 and maximum values based on what it analyzes should be masked. 

The test.py file runs everything, making a window in wxPython, plotting things with matplotlib, etc.

To switch between the mask and image, the code must be modified. An automated print function shows progress during running the program, as it can take some time on low RAM.